Fixed scrollbar with useState
